refactor(platform): address CodeRabbit review on the perf refactor
Verified each finding against current code; fixed the in-scope ones, skipped
pre-existing / out-of-scope items.

Fixed:
- Extract periodToDays + FeedbackPeriod into a pure util (feedback-period.ts) so
  the governance/feedback route loader no longer imports the UI page module.
- transform_conversation: only trust a prefetched customer whose _id matches the
  conversation (defensive); mismatch falls back to a direct fetch.
- file_metadata.listChatAttachmentsForThread: .collect() -> for-await stream.
- require_org_membership + lib/auth: getAuthUserIdentity now imported from its
  concrete module (the sibling-'../rls' barrel form the earlier sweep missed).
- paged_delete docstring: drop the contested "out-of-band" claim about
  storage.delete; state the DB-only intent neutrally.

Skipped (with reason): pre-existing security/quality findings outside this
perf PR's diff (share_thread org-auth, compare_documents team ACL, bulk
legal-hold OCC, notifications unread-count org guard, projects/threads cross-org
guards, deprecated create/deleteOrganization tombstones, agents/webhooks +
collab .collect()) — each warrants its own reviewed change. checkFileRagStatuses
returning null on no-auth and slug createdBy email-first are pre-existing
behaviors, not introduced here.
