+++ services/platform/convex/conversations/transform_conversation.ts
@@ -24,11 +24,21 @@ export async function transformConversation(
 ): Promise<ConversationItem> {
   const includeAllMessages = options?.includeAllMessages ?? false;
   const customerPrefetched = options !== undefined && 'customer' in options;
+  // Only trust a prefetched customer that actually belongs to this conversation
+  // (guards against a caller mapping the wrong customer); a prefetched `null` is
+  // trusted as "no customer". Anything else falls back to a direct fetch.
+  const prefetched = customerPrefetched
+    ? (options.customer ?? null)
+    : undefined;
+  const prefetchUsable =
+    prefetched === undefined ||
+    prefetched === null ||
+    prefetched._id === conversation.customerId;
 
   // Load customer and messages in parallel
   const [customerDoc, messageDocs] = await Promise.all([
-    customerPrefetched
-      ? Promise.resolve(options.customer ?? null)
+    customerPrefetched && prefetchUsable
+      ? Promise.resolve(prefetched ?? null)
       : conversation.customerId
         ? ctx.db.get(conversation.customerId)
         : Promise.resolve(null),
